==> 1_data_processing/01_process_chrom_info.ipynb <==
## Get coordinates of fragile sites, centromeres and telomeres
- Telomeres are all 10,000 bp (0.01 Mb) long
- Centromeres are all 3,000,000 bp (3Mb) long

import numpy as np
import pandas as pd
import scipy.stats as stats
import os

get_data_path = lambda folders, fname: os.path.normpath(os.environ['THIRD_PARTY_DIR']+'/'+'/'.join(folders)+'/'+fname)
get_local_data_path = lambda folders, fname: os.path.normpath('../local_data/' +'/'.join(folders) +'/'+ fname)

file_pcawg_fsites = get_data_path(['PCAWG'], 'major_fragile_sites.txt')
file_genome_gaps = get_data_path(['UCSC_genome'], 'hg19_gaps.txt')
file_chr_sizes = get_data_path(['UCSC_genome'], 'hg19.chrom.sizes.txt')

# Output
file_chrom_info = get_local_data_path(['processed'], 'hg19_chrom_info.csv')
file_fragsites = get_local_data_path(['processed'], 'hg19_fragile_sites.csv')

### Fragile sites

pcawg_fragsites = pd.read_csv(file_pcawg_fsites, sep=' ')
pcawg_fragsites = pcawg_fragsites[pcawg_fragsites.chrom!='chrX'].assign(chrom = lambda x: x.chrom.apply(lambda c: int(c[3:])))
# Update PARK2 symbol to PRKN (manually checked this in HGNC)
pcawg_fragsites = pcawg_fragsites.rename(columns={'start':'startpos','end':'endpos','chrom':'chr'}).replace({'PARK2':'PRKN'})
print('N major fragile sites in PCAWG (non-X):', pcawg_fragsites.shape[0])
pcawg_fragsites.to_csv(file_fragsites, index=0)

pcawg_fragsites

### Centromeres + telomeres

genome_gaps = pd.read_csv(file_genome_gaps, sep='\t')[['chrom','chromStart','chromEnd','type']]
genome_gaps = genome_gaps[genome_gaps.type.isin(['telomere','centromere'])]
genome_gaps['chrom'] = genome_gaps['chrom'].apply(lambda x: x[3:])
centromeres = genome_gaps[genome_gaps.type=='centromere'].reset_index(drop=True)
telomeres = genome_gaps[genome_gaps.type=='telomere'].reset_index(drop=True)

centromeres[:1]

chr_sizes = pd.read_csv(file_chr_sizes, sep='\t', nrows=24, names=['chrom','size'])
chr_sizes['chrom'] = chr_sizes['chrom'].apply(lambda x: x[3:])

assert(pd.merge(chr_sizes, telomeres.rename(columns={'chromEnd':'size'})).shape[0]==23) # chr 17 missing from telomeres

chrom_info = pd.merge(centromeres, chr_sizes).rename(columns={'chromStart':'centStart','chromEnd':'centEnd','chrom':'chr'})
chrom_info = chrom_info.drop(columns=['type']).set_index('chr')
chrom_info[:2]

chrom_info.to_csv(file_chrom_info)
