==> 1_data_processing/02_process_TCGA_ASCAT.ipynb <==
## TCGA ASCAT results

Data downloaded from: https://github.com/VanLoo-lab/ascat/tree/master/ReleasedData/TCGA_SNP6_hg19

These calls are all based on **hg19/GRCh37** coordinates.

From CCDS Readme:  
- cds_from = chromosome position of CDS start (or CDS end for minus strand genes), in 0-based coordinates      
- cds_to = chromosome position of CDS end (or CDS start for minus strand genes), in 0-based coordinates  
- cds_locations = comma separated list of from-to, of chromosome positions of CDS exons, in 0-based coordinates, ordered from low to high

import numpy as np
import pandas as pd
import scipy.stats as stats
import os
import matplotlib.pyplot as plt

import processing_utils as util
%load_ext autoreload
%autoreload 1
%aimport processing_utils

#get_data_path = lambda folders, fname: os.path.normpath(os.environ['THIRD_PARTY_DIR']+'/'+'/'.join(folders)+'/'+fname)
get_local_data_path = lambda folders, fname: os.path.normpath('../local_data/' +'/'.join(folders) +'/'+ fname)

folder_ascat_segments = get_local_data_path(['raw','TCGA','ASCATv3_SNP6_hg19'], 'segments')
file_ascat_summary = get_local_data_path(['raw','TCGA','ASCATv3_SNP6_hg19'],'summary.ascatv3TCGA.penalty70.hg19.tsv')
#file_ccds = get_data_path(['ccds_gene_annotations'], 'hg19.txt')
#file_chrom_info = get_local_data_path(['processed'], 'hg19_chrom_info.csv')

# OUTPUT
file_ascat_homdels = get_local_data_path(['processed','TCGA','ASCAT'], 'homdels.csv')
file_ascat_loh = get_local_data_path(['processed','TCGA','ASCAT'], 'LOH.csv')
file_ascat_segment_bounds = get_local_data_path(['processed','TCGA','ASCAT'], 'segment_boundaries.csv')

### ASCAT summary

ascat_summary_full = pd.read_csv(file_ascat_summary, sep='\t')
ascat_summary_full[:1]

ascat_summary = ascat_summary_full[(ascat_summary_full['QC']=='Pass')]
print('N samples with solution:', ascat_summary.shape[0],'/',ascat_summary_full.shape[0])
assert(ascat_summary.patient.nunique()==ascat_summary.shape[0])

### Extract homozygous deletion segments
All segments with total copy number 0.

homdels_list = []
for i, sname in enumerate(ascat_summary.name.values):
    print(i, end='\r')
    sample = pd.read_csv(folder_ascat_segments+'/'+sname+'.segments.txt', sep='\t')
    sample = sample.assign(CN = sample.nMajor + sample.nMinor)
    homdels_list.append(sample[sample.CN==0])

homdels_all = pd.concat(homdels_list).reset_index(drop=True)
homdels_all = pd.merge(ascat_summary[['name','patient','cancer_type','barcodeTumour']], 
                       homdels_all.rename(columns={'sample':'name'}))
homdels_all['del_len'] = homdels_all.endpos - homdels_all.startpos

homdels_all[:2]

# Check that number of HD segments matches summary report
print('N HDs:', homdels_all.shape[0])
qc_check = pd.merge(homdels_all.groupby('patient').size().reset_index(name='n_segs'), 
                    ascat_summary[['patient','homdel_segs']], how='right').fillna(0).astype({'n_segs':int})
assert(qc_check[qc_check.n_segs!=qc_check.homdel_segs].shape[0]==0)

homdels_all.drop(columns=['nMajor','nMinor','CN']).to_csv(file_ascat_homdels, index=0)

### Extract all LOH segments (minor cn = 0; major cn >0)

loh_list = []
for i, sname in enumerate(ascat_summary.name.values):
    print(i, end='\r')
    sample = pd.read_csv(folder_ascat_segments+'/'+sname+'.segments.txt', sep='\t')
    loh_list.append(sample[(sample.nMinor==0) & (sample.nMajor>0)])

loh = pd.concat(loh_list).reset_index(drop=True)

loh = pd.merge(ascat_summary[['name','patient','ploidy']], loh.rename(columns={'sample':'name'}))
loh = loh[~loh.chr.isin(['X','Y'])].astype({'chr':int}).reset_index(drop=True)
loh = loh.assign(del_len = loh.endpos - loh.startpos)

print('N LOH segments:', loh.shape[0])
print('N samples w/ LOH segments:', loh.patient.nunique())

loh.to_csv(file_ascat_loh, index=0)

### Identify ASCAT segment boundaries - telomere and centromere adjacent segments
- First segment start + last segment end for each chromosome.  
- Expect this to be the same for each sample, bounded by the SNP array, but check!   
- NOTE: short arm not mapped for chromosomes 13,14,15,21,22

chrom_info = pd.read_csv(file_chrom_info, index_col=0)
chrom_info[:1]

chr_start = []
chr_end = []
cent_start = []
cent_end = []
for i, sname in enumerate(ascat_summary.name.values):
    print(i, end='\r')
    sample = pd.read_csv(folder_ascat_segments+'/'+sname+'.segments.txt', sep='\t')
    # Find max endpos that is before centromere start, for each chr
    df = sample.groupby('chr').apply(lambda x: x.endpos[x.endpos <= chrom_info.loc[x.name,'centStart']].max())\
               .reset_index(name='endpos')
    cent_start.append(df)
    # Find min startpos that is after centromere end, for each chr
    df = sample.groupby('chr').apply(lambda x: x.startpos[x.startpos >= chrom_info.loc[x.name,'centEnd']].min())\
               .reset_index(name='startpos')
    cent_end.append(df)
    # Find edges - overall min startpos / max endpos
    chr_start.append(sample.groupby('chr').startpos.min().reset_index())
    chr_end.append(sample.groupby('chr').endpos.max().reset_index())

chr_start_posns = pd.concat(chr_start).reset_index(drop=True)
chr_end_posns = pd.concat(chr_end).reset_index(drop=True)
cent_start_posns = pd.concat(cent_start).reset_index(drop=True)
cent_end_posns = pd.concat(cent_end).reset_index(drop=True)

# Check whether start and end is the same for every sample
for chrom in range(1,23):
    print('Chr', chrom, '\tstart:', chr_start_posns[chr_start_posns.chr==str(chrom)].startpos.unique(),
          '   \tend:', chr_end_posns[chr_end_posns.chr==str(chrom)].endpos.unique())

bounds = pd.merge(chr_start_posns.groupby('chr').startpos.min().reset_index(), 
                  chr_end_posns.groupby('chr').endpos.max().reset_index())
# No segments before centromere on chromosomes 13, 14, 15, 22
cent_bounds = pd.merge(cent_start_posns.groupby('chr').endpos.max().reset_index().rename(columns={'endpos':'cent_start'}),
                       cent_end_posns.groupby('chr').startpos.min().reset_index().rename(columns={'startpos':'cent_end'}))
bounds = pd.merge(bounds, cent_bounds)
bounds = bounds[~bounds.chr.isin(['X','Y'])].astype({'chr':int}).sort_values('chr').set_index('chr')

bounds

bounds.to_csv(file_ascat_segment_bounds)
